+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Support for per-node resources
 - Number of CPU cores, number of GPUs, the amount of memory and the amount of storage can be now requested per-node using the submission options `ncpus-per-node`, `ngpus-per-node`, `mem-per-node`, and `work-size-per-node`. Per-node properties override per-cpu properties (`mem-per-cpu`, `work-size-per-cpu`) but are overriden by "total" properties (`ncpus`, `ngpus`, `mem`, `work-size`).
 
+### Bug fixes and minor improvements
+- The available types of working directories for the current environment are now shown in the output of `qq submit -h`.
+
 ***
 
 ## Version 0.5.1